+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/BooleanValuesTest.java
@@ -88,7 +88,7 @@ public class BooleanValuesTest {
 						.buildValue();
 
 		NullableCheckBox checkBox = (NullableCheckBox) value.component();
-		NullableToggleButtonModel model = checkBox.getNullableModel();
+		NullableToggleButtonModel model = checkBox.model();
 
 		assertNull(value.get());
 		model.setSelected(true);